Made tests and moved helper
