+++ src/utils/general-helpers.ts
@@ -86,6 +86,3 @@ export const setDelay = (ms: number) => new Promise((resolve) => setTimeout(reso
 const nbNoNumberFormat = new Intl.NumberFormat('nb-NO');
 
 export const formatLocaleNumber = (n: number) => nbNoNumberFormat.format(n);
-
-export const percentageOfAComparedToB = (numberA?: number, numberB?: number): number | undefined =>
-  numberA !== undefined && numberB !== undefined && numberB > 0 ? Math.round((numberA / numberB) * 100) : undefined;
